Create README.md
I have developed an interactive navigation menu using HTML, CSS and JavaScript as a 1st task in web development internship at PRODIGY . This navigation menu changes color or style when scrolled or when hovering over a menu item. The navigation menu should have a fixed position and be visible on all pages. I had use HTML to structure the menu, CSS to style it, and JavaScript to add interactivity, such as changing the background color or font color of the menu when it is scrolled or when a menu item is hovered over. Really It was an great experience after completing this task.